refactor: use parceljs for JS modules
