+++ extension/.prettierrc
@@ -1,0 +1,7 @@
+{
+	"useTabs": false,
+	"printWidth": 100,
+	"semi": true,
+	"singleQuote": true,
+	"trailingComma": "es5"
+}
