--- extension/popup.html
@@ -1,96 +1,0 @@
-<!--
-Allows user to adjust filtration settings.
- --> 
-<!DOCTYPE html>
-<html>
-<head>
-	<title>Options</title>
-	<style>
-	body {
-		width: 20rem;
-		margin: 0;
-	}
-	.container {
-		display: flex;
-		flex-direction: column;
-		padding: 1rem;
-		font-size: 1.25rem;
-		background-color: #f2f2ff;
-		color: #382381;
-		border: 2px #183578;
-	}
-	#title {
-		padding-top: 0.3rem;
-		padding-bottom: 0.7rem;
-		font-size: 1.75rem;
-		font-weight: bold;
-		text-align: center;
-	}
-	.option-row {
-		display: flex;
-		justify-content: space-between;
-		flex-direction: row;
-		margin-top: 0.5rem;
-	}
-	input[type=range] {
-		width: 100%;
-	}
-	input[type=checkbox] {
-		transform: scale(1.5);
-	}
-	select {
-		width: 50%;
-		font-size: 1.25rem;
-		text-align: center;
-		background-color: #f2f2ff;
-	}
-	</style>
-</head>
-<body>
-	<div class="container">
-		<div id="options">
-			<div id="title" >
-				<span>Options</span>
-			</div>
-			<div class="option-row" title="Choose how to alter negative&#013;items' visibility">
-				<label for="selected-styling">Styling mode</label>
-				<select id="selected-styling" class="stored-options">
-					<option value="opacity">&nbspTransparency&nbsp</option>
-					<option value="color">&nbspColors&nbsp</option>
-					<option value="remove">&nbspRemove&nbsp</option>
-				</select>
-			</div>
-
-			<div class="option-row" title="Increase to show more / Decrease for less">
-				<label for="selected-threshold">Negativity tolerance</label>
-				<input class="slider stored-options" id="selected-threshold"
-					   type="range" min="0" max="100" value="50">
-			</div>
-
-			<div class="option-row"
-				 title="Toggle to use relative sensitivity&#013;(e.g. set threshold to 20% to show only 20%)">
-				<label for="selected-ranking-check">Use percentage values</label>
-				<input class="stored-options" type="checkbox" id="selected-ranking-check">
-			</div>
-
-			<div class="option-row"
-				 title="Toggle to restyle only texts elements&#013; themselves and not their highest container">
-				<label for="selected-onlytexts-check">Only text elements</label>
-				<input class="stored-options" type="checkbox" id="selected-onlytexts-check">
-			</div>
-
-            <div class="option-row" title="How to calculate negativity values">
-				<label for="selected-backend">Calculation</label>
-				<select id="selected-backend" class="stored-options">
-					<option value="python">&nbspPython&nbsp</option>
-					<option value="simple">&nbspJS-Stub&nbsp</option>
-					<option value="off">&nbspOff&nbsp</option>
-				</select>
-			</div>
-
-		</div>
-	</div>
-	<script type="text/javascript" src="js/popup.js"></script>
-</body>
-</html>
-
