+++ extension/src/background.js
@@ -1,0 +1,1 @@
+console.log('hello from background');
