+++ extension/src/content-script.js
@@ -1,0 +1,1 @@
+import './content/main';
